Adds nodevisitor that makes all named nodes always defined.
This fixes an issue where gettext can not correctly find plural strings.

See https://github.com/php-gettext/Gettext/issues/229
