+++ clarkson-wp-cli-twig-translations.php
@@ -9,10 +9,7 @@ License: GPL 2.0
 
 namespace Clarkson\WPCLITwigTranslations;
 
-use Twig\Extra\Html\HtmlExtension;
-use Twig\Extra\Intl\IntlExtension;
-use Twig\Extra\Markdown\MarkdownExtension;
-use Twig\Extra\String\StringExtension;
+use WP_CLI;
 
 if( !defined('WP_CLI') || ! WP_CLI ) {
     return;
@@ -23,46 +20,8 @@ if ( file_exists( $autoload ) ) {
 	require_once $autoload;
 }
 
-if( ! class_exists( '\Clarkson\WPCLITwigTranslations\Gettext' ) ){
-    class Gettext extends \WP_CLI_Command {
-        public function __invoke( $args, $assoc_args ) {
-            $twig_args = array(
-                'debug' => true,
-                'cache' => __DIR__ . '/../../../' . $assoc_args['output-dir'] ,
-                'auto_reload' => true
-            );
-
-            $basedir = realpath( __DIR__ . '/../../../' . $assoc_args['template-dir'] );
-            $twig_fs = new \Twig\Loader\FilesystemLoader( $basedir );
-            $twig 	 = new \Twig\Environment( $twig_fs, $twig_args );
-            $twig->registerUndefinedFunctionCallback(function ($name) {
-                return new \Twig\TwigFunction($name, $name);
-            });
-            $twig->addExtension( new IntlExtension() );
-            $twig->addExtension( new StringExtension() );
-            $twig->addExtension( new HtmlExtension() );
-            $twig->addExtension( new MarkdownExtension() );
-
-            $filelist = $this->getFileList( $basedir );
-
-            foreach($filelist as $filepath) {
-                $truepath = str_replace( $basedir, '', $filepath );
-                $twig->load( $truepath );
-            }
-        }
-
-        protected function getFileList( $dir ) {
-            $filelist = array();
-            $directories = glob( realpath( $dir ). '/*', GLOB_ONLYDIR );
-            foreach( $directories as $directory ){
-                $filelist = array_merge( $filelist, $this->getFileList( $directory ) );
-            }
-            $filelist = array_merge( $filelist, glob( realpath( $dir ) . '/*.twig' ) );
-            return $filelist;
-        }
-    }
-
-    \WP_CLI::add_command('clarkson-twig-translations', "\Clarkson\WPCLITwigTranslations\Gettext", array(
+if( ! class_exists( GetText::class ) ){
+    WP_CLI::add_command('clarkson-twig-translations', GetText::class, array(
         'synopsis' => array(
             array(
                 'type'        => 'assoc',
